move generate schema code to `odra-build` crate
