+++ odra-build/Cargo.toml
@@ -7,3 +7,7 @@ authors.workspace = true
 license.workspace = true
 homepage.workspace = true
 repository.workspace = true
+
+[dependencies]
+serde = { workspace = true }
+serde_json = { workspace = true, default-features = false, features = ["alloc"] }
